Shuffle functionality that takes a deck and shuffles it. Since Go doesn't have a randomizer a customer number randomizer had to  be created.
using the time I can get the exact moment and turn it to a UnixNano() that returns an int64. This is used as my source which when passed using the rand standard library in Go, I can create a new instance which allows me to also use Intn - Intn returns, as an int, a non-negative pseudo-random number in the half-open interval [0,n).

This is then used as the 'new position' for the deck, and we reassign the values of the passed deck with the random integer created as the key instead of the keys that were passed when the deck function initilized.